feature:24.11.2023 17:51 in class AuthorServiceImpl was rewritten methods getByNameV1, getByNameV2, getByNameV3, createAuthor, updateAuthor, deleteAuthor and getAllAuthors, new method getAuthorByNameSpecification was added.

+++ library-project/src/main/java/ru/babushkina/libraryproject/service/AuthorServiceImpl.java
@@ -90,27 +90,54 @@ public class AuthorServiceImpl implements AuthorService{
 
     @Override
     public AuthorDto getByNameV2(String name) {
-        Author author = authorRepository.findAuthorByNameBySQL(name).orElseThrow();
-        return convertEntityToDto(author);
+        log.info("Get author by name: {}", name);
+        Optional<Author> author = authorRepository.findAuthorByName(name);
+        if (author.isPresent()) {
+            AuthorDto authorDto = convertEntityToDto(author.get());
+            log.info("Author: {}", authorDto.toString());
+            return authorDto;
+        } else {
+            log.error("Author with name {} not found", name);
+            throw new NoSuchElementException("No value present");
+        }
     }
 
     @Override
     public AuthorDto getByNameV3(String name) {
-        Specification<Author> specification = Specification.where(new Specification<Author>() {
-            @Override
-            public Predicate toPredicate(Root<Author> root, CriteriaQuery<?> query, CriteriaBuilder criteriaBuilder) {
-                return criteriaBuilder.equal(root.get("name"), name);
-            }
+        log.info("Get author by name: {}", name);
+        Specification<Author> specification = getAuthorByNameSpecification(name);
+        Author author = authorRepository.findOne(specification).orElseThrow(() -> {
+            log.error("Author with name {} not found", name);
+            return new NoSuchElementException("No value present");
         });
-        Author author = authorRepository.findOne(specification).orElseThrow();
-        return convertEntityToDto(author);
+        AuthorDto authorDto = convertEntityToDto(author);
+        log.info("Author: {}", authorDto.toString());
+        return authorDto;
+//        Specification<Author> specification = Specification.where(new Specification<Author>() {
+//            @Override
+//            public Predicate toPredicate(Root<Author> root, CriteriaQuery<?> query, CriteriaBuilder criteriaBuilder) {
+//                return criteriaBuilder.equal(root.get("name"), name);
+//            }
+//        });
+//        Author author = authorRepository.findOne(specification).orElseThrow();
+//        return convertEntityToDto(author);
+    }
+
+    private Specification<Author> getAuthorByNameSpecification(String name) {
+        return ((root, query, criteriaBuilder) -> criteriaBuilder.equal(root.get("name"), name));
     }
 
     @Override
     public AuthorDto createAuthor(AuthorCreateDto authorCreateDto) {
-        Author author = authorRepository.save(convertDtoToEntity(authorCreateDto));
+        log.info("Info about author: {}", authorCreateDto.toString());
+        Author author = convertDtoToEntity(authorCreateDto);
+        author = authorRepository.save(author);
         AuthorDto authorDto = convertEntityToDto(author);
+        log.info("Author created: {}", authorDto.toString());
         return authorDto;
+//        Author author = authorRepository.save(convertDtoToEntity(authorCreateDto));
+//        AuthorDto authorDto = convertEntityToDto(author);
+//        return authorDto;
     }
 
     private Author convertDtoToEntity(AuthorCreateDto authorCreateDto) {
@@ -144,22 +171,49 @@ public class AuthorServiceImpl implements AuthorService{
 
     @Override
     public AuthorDto updateAuthor(AuthorUpdateDto authorUpdateDto) {
-        Author author = authorRepository.findById(authorUpdateDto.getId()).orElseThrow();
+        log.info("Update author: {}", authorUpdateDto.toString());
+        Author author = authorRepository.findById(authorUpdateDto.getId()).orElseThrow(() -> {
+            log.error("Author with id {} not found", authorUpdateDto.getId());
+            return new NoSuchElementException("No value present");
+        });
         author.setName(authorUpdateDto.getName());
         author.setSurname(authorUpdateDto.getSurname());
         Author savedAuthor = authorRepository.save(author);
         AuthorDto authorDto = convertEntityToDto(savedAuthor);
+        log.info("Author updated: {}", authorDto.toString());
         return authorDto;
+//        Author author = authorRepository.findById(authorUpdateDto.getId()).orElseThrow();
+//        author.setName(authorUpdateDto.getName());
+//        author.setSurname(authorUpdateDto.getSurname());
+//        Author savedAuthor = authorRepository.save(author);
+//        AuthorDto authorDto = convertEntityToDto(savedAuthor);
+//        return authorDto;
     }
 
     @Override
     public void deleteAuthor(Long id) {
-        authorRepository.deleteById(id);
+        log.info("Delete author with id: {}", id);
+        Optional<Author> author = authorRepository.findById(id);
+        if (author.isPresent()) {
+            authorRepository.deleteById(id);
+            log.info("Author with id {} was deleted", id);
+        } else {
+            log.error("An error acquired while deleting author with id {}", id);
+            throw new NoSuchElementException("No value present");
+        }
+//        authorRepository.deleteById(id);
     }
 
     @Override
     public List<AuthorDto> getAllAuthors() {
+        log.info("Get all authord");
         List<Author> authors = authorRepository.findAll();
-        return authors.stream().map(this::convertEntityToDto).collect(Collectors.toList());
+        List<AuthorDto> authorDtos = authors.stream()
+                .map(this:: convertEntityToDto)
+                .collect(Collectors.toList());
+        log.info("Found {} authors", authorDtos.size());
+        return authorDtos;
+//        List<Author> authors = authorRepository.findAll();
+//        return authors.stream().map(this::convertEntityToDto).collect(Collectors.toList());
     }
 }
